build: make plist/imobiledevice lib check more useful
The current checks for plist and imobiledevice were a bit useless. If
they were found, nothing was done. And if they weren't, nothing was done
either. We could just drop the tests since we should be able to trust
`PKG_CHECK_MODULES`.

As an alternative, this commit triggers a failure if the link is
unsuccessful. It also ensures that linking can be done when the libs are
located in some unstandard directory.